Use double in place of float for boundaries

+++ src/gui/elems/sampleEditor/waveform.cpp
@@ -145,13 +145,13 @@ int geWaveform::alloc(int datasize, bool force)
 		
 		/* Scan the original waveform in chunks [pc, pn]. */
 
-		float pc = i     * m_ratio;  // current point
-		float pn = (i+1) * m_ratio;  // next point
+		double pc = i     * m_ratio;  // current point
+		double pn = (i+1) * m_ratio;  // next point
 
 		float peaksup = 0.0f;
 		float peakinf = 0.0f;
 
-		for (int k=pc; k<pn; k++) {
+		for (int k=pc; k<pn; k++) { // it's ok to use int here. TODO - until we switch to uint32_t for Wave size...
 
 			if (k >= wave->getSize())
 				continue;
